Changed Doubles and doubles to BigDecimal

+++ src/main/java/pique/evaluation/IEvaluator.java
@@ -22,6 +22,8 @@
  */
 package pique.evaluation;
 
+import java.math.BigDecimal;
+
 import pique.model.ModelNode;
 
 // TODO (1.0): documentation
@@ -31,7 +33,7 @@ public interface IEvaluator {
      * @return
      *      The evaluated value of a node after running evalation given that node's state
      */
-    double evaluate(ModelNode inNode);
+    BigDecimal evaluate(ModelNode inNode);
 
     /**
      * @return
