+++ src/main/java/pique/evaluation/LoCEvaluator.java
@@ -26,6 +26,7 @@ import pique.model.Finding;
 import pique.model.ModelNode;
 import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -34,7 +35,7 @@ public class LoCEvaluator extends Evaluator {
 
     // TODO (1.0): Documentation
     @Override
-    public double evaluate(ModelNode inNode) {
+    public BigDecimal evaluate(ModelNode inNode) {
 
         // TODO (1.0): Add error handling
         throw new NotImplementedException();
