+++ src/main/java/pique/model/Tqi.java
@@ -24,6 +24,7 @@ package pique.model;
 
 import pique.evaluation.*;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -38,19 +39,19 @@ public class Tqi extends ModelNode {
 
     // Constructor
 
-    public Tqi(String name, String description, Map<String, Double> weights) {
+    public Tqi(String name, String description, Map<String, BigDecimal> weights) {
         super(name, description, new DefaultFactorEvaluator(), new DefaultNormalizer());
         this.weights = (weights == null) ? new HashMap<>() : weights;
     }
 
     public Tqi(String name, String description, IEvaluator evaluator, INormalizer normalizer,
-               IUtilityFunction utilityFunction, Map<String, Double> weights, Double[] thresholds) {
+               IUtilityFunction utilityFunction, Map<String, BigDecimal> weights, BigDecimal[] thresholds) {
         super(name, description, evaluator, normalizer, utilityFunction, weights, thresholds);
     }
 
     // Used for cloning
-    public Tqi(double value, String name, String description, IEvaluator evaluator, INormalizer normalizer,
-               IUtilityFunction utilityFunction, Map<String, Double> weights, Double[] thresholds, Map<String,
+    public Tqi(BigDecimal value, String name, String description, IEvaluator evaluator, INormalizer normalizer,
+               IUtilityFunction utilityFunction, Map<String, BigDecimal> weights, BigDecimal[] thresholds, Map<String,
             ModelNode> children) {
         super(value, name, description, evaluator, normalizer, utilityFunction, weights, thresholds, children);
     }
